add validation to deleteAuth method (no null or empty authtokens may be deleted). Add negative test case.

+++ server/src/main/java/dataaccess/MySQLAuthDAO.java
@@ -87,6 +87,9 @@ public class MySQLAuthDAO implements AuthDAO{
 
     @Override
     public void deleteAuth(String authToken) throws DataAccessException {
+        if(authToken == null || authToken.isEmpty()){
+            throw new DataAccessException("Invalid authToken, cannot delete.");
+        }
         try (var conn = DatabaseManager.getConnection()) {
             var statement = """
                 DELETE FROM auth
